♻️ Implement AuthController for registration and login #35
