--- src/Controller/LoginController.php
@@ -1,34 +1,0 @@
-<?php
-
-namespace App\Controller;
-
-use App\core\Controller;
-use App\core\Validation\Validator;
-use App\Manager\UserManager;
-
-class LoginController extends Controller
-{
-    public function login()
-    {
-        $post = $this->post;
-        $userManager = new UserManager();
-        if (!empty($post)) {
-            $validations = new Validator($post);
-            $errors = $validations->validate('login');
-            if (empty($errors)) {
-                $result = $userManager->login($post['username'], $post['password']);
-                if (true === $result) {
-                    header('Location: /');
-
-                    exit;
-                }
-                $errors['login'] = 'Le couple " Nom d\'utilisateur " et " Mot de passe " ne correspondent pas. ';
-            }
-        }
-
-        return $this->render('login/index.twig', [
-            'errors' => $errors ?? null,
-            'post' => $post ?? null,
-        ]);
-    }
-}
